+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -460,7 +460,7 @@ public class EntityPanel extends JPanel {
   /**
    * @param <T> the edit panel type
    * @return the edit panel
-   * @throws IllegalStateException in case no edit panel is avilable
+   * @throws IllegalStateException in case no edit panel is available
    * @see #containsEditPanel()
    */
   public final <T extends EntityEditPanel> T editPanel() {
@@ -481,7 +481,7 @@ public class EntityPanel extends JPanel {
   /**
    * @param <T> the table panel type
    * @return the table panel
-   * @throws IllegalStateException in case no table panel is avilable
+   * @throws IllegalStateException in case no table panel is available
    * @see #containsTablePanel()
    */
   public final <T extends EntityTablePanel> T tablePanel() {
@@ -776,6 +776,14 @@ public class EntityPanel extends JPanel {
   }
 
   /**
+   * Returns the base panel containing the edit panel and controls.
+   * @return the edit base panel
+   */
+  protected final JPanel editControlPanel() {
+    return editControlPanel;
+  }
+
+  /**
    * Returns the base panel containing the edit and table panels (north, center).
    * @return the edit and table base panel
    */
@@ -944,7 +952,6 @@ public class EntityPanel extends JPanel {
   protected final void initializeEditPanel() {
     if (editPanel != null) {
       editPanel.initialize();
-      editControlPanel.add(createEditBasePanel(editPanel), BorderLayout.CENTER);
       if (configuration.includeControls) {
         JComponent controlComponent = createControlComponent(createControls());
         if (controlComponent != null) {
@@ -984,14 +991,18 @@ public class EntityPanel extends JPanel {
       return null;
     }
 
-    JPanel panel = new JPanel(new BorderLayout());
-    panel.add(tablePanel, BorderLayout.CENTER);
-
-    return panel;
+    return new JPanel(new BorderLayout());
   }
 
   private JComponent mainComponent() {
-    return detailPanels.isEmpty() ? editControlTablePanel : detailLayout().layout(this);
+    if (editPanel != null) {
+      editControlPanel.add(createEditBasePanel(editPanel), BorderLayout.CENTER);
+    }
+    if (tablePanel != null) {
+      editControlTablePanel.add(tablePanel, BorderLayout.CENTER);
+    }
+
+    return detailLayout().layout(this);
   }
 
   final void setParentPanel(EntityPanel parentPanel) {
@@ -1349,6 +1360,7 @@ public class EntityPanel extends JPanel {
 
     /**
      * Creates and lays out the component to use as the main component of the given entity panel, including its detail panels.
+     * In case of no detail panels, this method should return the {@link EntityPanel#editControlTablePanel()}.
      * @param entityPanel the panel to lay out and configure
      * @return the main component
      */